+++ variable/CMakeLists.txt
@@ -12,7 +12,6 @@ set(INC_FILES
     include/scipp/variable/except.h
     include/scipp/variable/logical.h
     include/scipp/variable/math.h
-    include/scipp/variable/matrix.h
     include/scipp/variable/misc_operations.h
     include/scipp/variable/operations.h
     include/scipp/variable/rebin.h
@@ -22,6 +21,7 @@ set(INC_FILES
     include/scipp/variable/sort.h
     include/scipp/variable/special_values.h
     include/scipp/variable/string.h
+    include/scipp/variable/structures.h
     include/scipp/variable/subspan_view.h
     include/scipp/variable/transform.h
     include/scipp/variable/transform_subspan.h
@@ -47,7 +47,6 @@ set(SRC_FILES
     except.cpp
     logical.cpp
     math.cpp
-    matrix.cpp
     operations.cpp
     rebin.cpp
     reduction.cpp
@@ -56,6 +55,7 @@ set(SRC_FILES
     sort.cpp
     special_values.cpp
     string.cpp
+    structures.cpp
     subspan_view.cpp
     to_unit.cpp
     trigonometry.cpp
