+++ variable/include/scipp/variable/bin_array_model.h
@@ -10,7 +10,7 @@
 #include "scipp/core/except.h"
 #include "scipp/variable/arithmetic.h"
 #include "scipp/variable/cumulative.h"
-#include "scipp/variable/data_model.h"
+#include "scipp/variable/element_array_model.h"
 #include "scipp/variable/except.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/util.h"
