+++ variable/include/scipp/variable/element_array_model.h
