+++ variable/include/scipp/variable/structures.h
@@ -9,7 +9,7 @@
 namespace scipp::variable {
 
 template <class T, class Elem, int... N>
-[[nodiscard]] Variable make_structured(const Dimensions &dims,
+[[nodiscard]] Variable make_structures(const Dimensions &dims,
                                        const units::Unit &unit,
                                        element_array<double> &&values);
