+++ variable/test/CMakeLists.txt
@@ -6,8 +6,8 @@ set(TARGET_NAME "scipp-variable-test")
 add_dependencies(all-tests ${TARGET_NAME})
 add_executable(
   ${TARGET_NAME} EXCLUDE_FROM_ALL
+  bin_array_model_test.cpp
   bin_util_test.cpp
-  bucket_model_test.cpp
   comparison_test.cpp
   copy_test.cpp
   creation_test.cpp
